+++ src/Render/shader/gbuffer_frag.glsl
@@ -10,7 +10,7 @@ uniform int MtlIdx;
 layout (location=0)out vec3 Position;
 layout (location=1)out vec3 Normal;
 layout (location=2)out vec3 Texcoords;
-layout (location=3)out ivec3 FragIndex;
+layout (location=3)out vec3 FragIndex;
 
 void main()
 {
@@ -18,5 +18,5 @@ void main()
 	Normal=normalize(vNormal);
 	Texcoords=vec3(vTexcoords,1.f);
 	
-	FragIndex=ivec3(MeshIdx,MtlIdx,1);
+	FragIndex=vec3(MeshIdx,MtlIdx,1);
 }
